image-model-web/requirements.txt
flask
tensorflow
numpy
pillow 